theoretically.... split the reduced model, works faster fixed the reduced model abstraction.. enforced it with a bad-ass billy club moved highlighting to the reduced model
there seems to be a problem with the cursor... it doesn't show up any more
we don't know why, but the only thing that's been changed since the cursor
last worked was a test file.  so we don't think it's our fault.. we think
that the problem will just go away on it's own.


git-svn-id: file:///tmp/test-svn/trunk@69 fe72c1cf-3628-48e9-8b72-1c46755d3cff