1) Undo/redo are back! 2) System.out goes to interactions too (in green). We can debate this feature later,    and also we should make it configurable. 3) Fixed race condition after aborting interactions that sometimes caused    the interactions window to not have the correct classpath after an abort.
Note: A problem I see now but have not yet fixed is that indent is considered for undo
as some insertions and removals. It really should be considered as one atomic event.
Doing this would require a bunch of messing with how the undo/redo events are generated,
and it's not trivial. This is why I didn't implement this. We should do it at some
point, though.


git-svn-id: file:///tmp/test-svn/trunk@385 fe72c1cf-3628-48e9-8b72-1c46755d3cff