Added three more question nodes to the indentation tree with tests: QuestionLineContains, QuestionStartingNewStmt and QuestionExistsCharInStmt (previously known as QuestionInTernary).
Added a helper method to DefinitionsDocument called findCharOnLine(int pos,
char findChar). It searches for the given character on the line where the given
position is. The code for this helper was previously in QuestionLineContains,
but was abstracted out to this method. The testing for this function still
resides in QuestionLineContainsTest, however, and should probably be moved to
IndentHelperTest.


git-svn-id: file:///tmp/test-svn/trunk@543 fe72c1cf-3628-48e9-8b72-1c46755d3cff