Fixed a potential typing bug --
an OptionMap should not get an Enumeration<Option<Object>> but an
Enumeration<OptionParser<Object>> since OptionParser<Object> only
promises that the parsing result be an object.  An Option<T> is not
really a subclass of Option<Object> since the format(T val) method
accepts input of type T, not of type object.


git-svn-id: file:///tmp/test-svn/trunk@565 fe72c1cf-3628-48e9-8b72-1c46755d3cff