Improved the Display All Breakpoints message.
Synchronized the thread suspend and resume notifiers on DebugManager
so that the menu items aren't incorrectly shown and then hidden.

Added a message about how to use the debugger if tools.jar isn't
already on the classpath (on the debugger tab of preferences).

Changed checkbox preferences to have a label on the right, and
put a border around the toolbar icons button group.


git-svn-id: file:///tmp/test-svn/trunk@963 fe72c1cf-3628-48e9-8b72-1c46755d3cff