Resolved many threading issues between the debugger and Swing.  The Swing tutorial states that ONLY the event-dispatching thread should update the GUI, so several calls to SwingUtilities.invokeLater are now made to improve the behavior.
Also worked on the DebugPanel UI, added ability to add/remove watches in
the table, and gave the DebugPanel a DebugListener to decouple it from
MainFrame.

Added a HelpFrame and Help icon for browsing the user documentation.
The docs should be compiled on a machine with docbook using "ant docs".


git-svn-id: file:///tmp/test-svn/trunk@976 fe72c1cf-3628-48e9-8b72-1c46755d3cff