Improves the performance of the Debugger's step function. If duplicate step requests are encountered while the thread is suspended, the first step request is deleted.
Showing the threads tab is a configurable option now.

The values of watchpoints are updated as the watches are entered as well as
during each suspension of the thread.

The UncaughtExceptionWindow's layout has been changed. The ok button is now
at the bottom of the window instead of at the right side since this was
causing some text to be cut off.


git-svn-id: file:///tmp/test-svn/trunk@984 fe72c1cf-3628-48e9-8b72-1c46755d3cff