Improved config option component usability by adding more noticable buttons for font/color/keybinding choosers.
Fixed support for Swing in the Interactions pane on OS X by removing
the "headless AWT" option from the JVM startup.

Made background, error, and debug highlight colors configurable.

Removed old highlight tests from Definitions Pane.

Added instructions on generating docs in HelpFrame when in the
DEVELOPMENT mode.

Made Compile All more of the default (over Compile Current) and gave
it a key binding (F5).

Added "docs" to the release target in ant.

Started work on Revert All, but it is currently causing problems
with the reduced model when a doc is not displayed.


git-svn-id: file:///tmp/test-svn/trunk@997 fe72c1cf-3628-48e9-8b72-1c46755d3cff