This commit is mainly infrastructure changes which were necessary to start supporting JSR-14 v1.2 and Java 1.4.1, which have a different compiler interface.
There are now new source trees at the root level for classes which
can only be compiled with JSR-14 v1.0 or JSR-14 v1.2, which includes
many of the compiler adapters.  In this commit, JavacGJCompiler and
JSR14Compiler have been moved to the src/src-jsr14v1_0/ source tree,
and their pre-compiled classes have been added to compilers-jsr14v1_0.jar
in the lib directory, which should be added to your classpath.

With this commit, DrJava can now be compiled with either JSR-14 v1.0
or JSR-14 v1.2.  However, integrated support for the v1.2 compiler
within DrJava (via a compiler adapter) has not been added yet.


git-svn-id: file:///tmp/test-svn/trunk@1066 fe72c1cf-3628-48e9-8b72-1c46755d3cff