- Fixed a bug with undoing, where pressing <Enter> and then undoing would only undo the indentation of the new line, then another undo must be signaled in order to undo the insertion of the newline.
- Refactored OurUndoManager.  It was a private inner class on
DefinitionsDocument, but we moved it into its own public class
CompoundUndoManager so that the ui can begin compound edits.

- added functionality to the compound edit capacity of CompoundUndoManager
so that multiple CompoundEdits can be nested.  In order to enforce that
the edits are well-nested, we return a key (an int) when the edit is added
to the CompoundUndoManager, which must be passed into the
endCompoundEdit() method.


git-svn-id: file:///tmp/test-svn/trunk@1235 fe72c1cf-3628-48e9-8b72-1c46755d3cff