- Updated DependenciesTest to require the new InputStreamRedirector from util.
- Pulled functionality from InteractionsDocument into a ConsoleDocument
to allow input to be entered into the console at a prompt.

- MainJVM now asks for input from the DefaultInteractionsModel, which
redirects through DefaultGlobalModel to an InputListener which must be
registered there.

- Moved listenerFail() method from GlobalModelTestCase to
MulitThreadedTestCase since it makes sense in any multithreaded test.

- Modified the interactions pane so that you can't edit/perform actions
while there is no prompt.  It is now considered to be inProgress while
resetting.

- Added a ConsoleController, with the code it shares with
InteractionsController moved to AbstractConsoleController.

- Added a constructor to InteractionsPane so that it can be used as the
console pane as well, and removed OutputPane.

- Added a right-click context menu to the document list pane.


git-svn-id: file:///tmp/test-svn/trunk@1397 fe72c1cf-3628-48e9-8b72-1c46755d3cff