- Fixed a bug in the debugger where if the thread tab was not enabled, we would get NullPointerExceptions all over the place.
- Fixed a bug where stepping into a breakpoint would push the same thread
onto the stack twice, and cause extremely unstable behavior in the
debugger, including the death of the EventHandler thread.

- If there was an unexpected expection in the EventHandler, it would not
catch it, and die.  As a result, the DebugPanel would not be hidden when
the interactions pane was reset.  Now, the EventHandler catches any
unexpected exceptions, and prints an error.  The debugger then continues
running so that the user can save any needed state before resetting.


git-svn-id: file:///tmp/test-svn/trunk@1401 fe72c1cf-3628-48e9-8b72-1c46755d3cff