Added findPrevious button which replaces the SearchBackwards CheckBox by actually executing a backwards search. The Enter key executes the behavior specified by the last pushed button (findNext or findPrevious). It is findNext by default.
Changed commenting code in DefinitionsPane to use the code written in MainFrame instead of code duplication.

Fixed a few bugs that were added into Find/Replace and fixed a long-standing bug that messed up the focus-switching in the Find/Replace Dialog when Replace All button was not active (i.e. when the _findField was blank).


git-svn-id: file:///tmp/test-svn/trunk@3281 fe72c1cf-3628-48e9-8b72-1c46755d3cff