Revised:
(i) the constructor in the ToolsJarClassLoader class to take a File
argument specifying the location of the tools.jar file stored in the
user's preferences (.drjava file).

(ii) the DrJava class to pass the stored location of tools.jar to the
ToolsJarClassLoader used to check if tools.jar can be found.

Effect: these revisions fix a bug in the startup protocol that displays a dialog box asking for the location of tools.jar when the information is already stored in the user's profile (.drjava file).


git-svn-id: file:///tmp/test-svn/trunk@3538 fe72c1cf-3628-48e9-8b72-1c46755d3cff