This commit
1. Improves the synchronization in FindReplaceMachine.  In the new
version of FindReplaceMachine, I dropped setting a modifyLock at the
beginning of _replaceAllInCurrentDoc because the unit tests would hang
at approxsimately this point.  I erroneously thought it was a flaw in
the implementation of readers/writers locking in
javax.swing.text.AbstractDocument.  In fact, the problem was a bug
(corrected in a very recent commit) in my code for
_replaceAllInCurrentDoc.  In both calls on FindNext that are confined
to the current document, I failed to pass (false) as an argument.  The
0-ary version uses the value of _searchAllDocuments for this argument
which is true in a global replace operation.  Hence, the modifyLock
that I was setting spanned documents.  (I don't excatly where it was
hanging up, but the corresponding modifyUnlock did not generally
unlock the same document!). That modifyLock has been reinstalled and
the remaining synchronization has been thought through a bit more
carefully.  It attempts to make each transaction on a document atomic.

2. Rationalized the naming of predicate methods and fields.  The "is"
prefix is used with all participles (e.g., "Modified", "Selected",
etc.) and phrases that do not necessarily sound like predicates
without the "is" prefix, but it explicitly omits the "is" prefix if
the next word is a preposition, since these phrases are clearly
predicates with the "is" prefix. (e.g., "onMatch").



git-svn-id: file:///tmp/test-svn/trunk@3785 fe72c1cf-3628-48e9-8b72-1c46755d3cff