Merged revisions 3734 through 3788 of the drjava-asynctask branch into the trunk.
These revisions include a few important bug fixes related to the interactions pane and definitions pane. This new build includes the async framework that will be used in the model to perform async operations with progress monitoring. Currently, only the "Clean Build 
Directory" is performed asynchronously, but the async opening of multiple files will be comming soon

git-svn-id: file:///tmp/test-svn/trunk@3808 fe72c1cf-3628-48e9-8b72-1c46755d3cff