This change fixes bug report [ 1561534 ] Debug Stepping Over not working with keyboard input
This was not a debugger error. The problem was the way the input box
was inserted into the interactions document, as I already assumed. Due
to the different prompt during debugging, it was inserted in an
invalid place, causing it not to show up and give the user a chance to
hit enter. There was nothing wrong with the debugger.

To insert the component, which is represented by a plain-old string
("[component]" in this case), our own document class was bypassed and
the Swing document accessed directly, probably because creating it was
inconvenient at the time to create a proper style string for the input
box and let our own document class know about it.

As of this revision, the input box will appear and stepping/resuming
will work. When the program finishes, sometimes there are two "> >"
symbols appearing now, so I suggest someone with better knowledge of
the parts of the program that contained the bug should take another
glance at the way the prompts are printed.

I removed the two files edu/rice/cs/util/swing/InputBox.java and
src/edu/rice/cs/util/swing/PopupConsole.java as they are not used
anymore and were red herrings during my initial attempts to locate the
bug. Also removed a retroweaver problem that had to do with my
previous commit for the memory statistics. 

M    src/edu/rice/cs/drjava/ui/InteractionsController.java
D    src/edu/rice/cs/util/swing/PopupConsole.java
D    src/edu/rice/cs/util/swing/InputBox.java
M    src/edu/rice/cs/util/StringOps.java


git-svn-id: file:///tmp/test-svn/trunk@4010 fe72c1cf-3628-48e9-8b72-1c46755d3cff