Fixes bug [ 1696060 ] Debugger Infinite Loop.
The deadlock was a result of the debugger thread, already owning the
lock of the debugger, asking the global model to open a file, so
setActiveDocument put code on the event queue. When this code
executed, the event thread tried to call a synchronized method of the
debugger, but never managed to acquire it, because the debugger thread
already owned it.

I changed the code to have the debugger thread
"preload" a document every time a step is made, before claiming the
debugger lock. This brings setActiveDocument into action and the
document is opened if necessary. Then the debugger lock is claimed and
the response to the step is made.

M    src/edu/rice/cs/drjava/model/debug/jpda/JPDADebugger.java
M    src/edu/rice/cs/drjava/model/debug/jpda/EventHandlerThread.java


git-svn-id: file:///tmp/test-svn/trunk@4225 fe72c1cf-3628-48e9-8b72-1c46755d3cff