Fixed a problem with loading projects and restoring the selection and active document
Completed the "project auto-refresh" task from COMP 312 that wasn't finished; now there is a menu item to invoke the feature on demand, and the black list can be edited in the "Project Properties" dialog.

Added several action properties that can be invoked from the command line of an external process. The actions are implemented as side effects of the evaluation of the command line: A variable used to denote an action performs the action and then evaluates to the empty string, disappearing from the command line.

So far, the following properties are available:

${action.save.all}
${action.compile.all}
${action.clean}
${action.open.file;file="Foo.java";line="100"}
${action.auto.refresh}


git-svn-id: file:///tmp/test-svn/trunk@4479 fe72c1cf-3628-48e9-8b72-1c46755d3cff