This commit allows the user to double-click on a word in the "Execute External Process" pane, and if it happens to be a file name, perhaps followed by a colon and a line number, then the definitions pane will display that file and jump to the specified line number. This helps with compilng files using an external compiler, such as javac, NextGen or xajavac, and then correcting the errors reported by the compilers.
A change was made to the way file names are saved in
project files: Even though we created a file name that was
relative to the project root, because we used the File
class, this file name was turned into an absolute path
in the project file. Now file names are treated as strings
and are now finally relative to the project root, and
not absolute.

Added catch statements to deal with a StringIndexOutOfBoundsException
[ 1967750 ] Some error with something....?
https://sourceforge.net/tracker/?func=detail&atid=438935&aid=1967750&group_id=44253



git-svn-id: file:///tmp/test-svn/trunk@4483 fe72c1cf-3628-48e9-8b72-1c46755d3cff