Fixes the bug [ 1992497 ] Folders and Bins Cannot Be Selected http://sourceforge.net/tracker/index.php?func=detail&aid=1992497&group_id=44253&atid=438935
Folders and bins now can be selected. If a folder is closed, the
selection in the navigator is moved to the document shown in the
Definitions Pane, thus making sure that something is selected in
the navigator, and that the selection is visible and not in a
collapsed subtree.

M    src/edu/rice/cs/drjava/ui/MainFrame.java
M    src/edu/rice/cs/util/docnavigation/JTreeSortNavigator.java


git-svn-id: file:///tmp/test-svn/trunk@4514 fe72c1cf-3628-48e9-8b72-1c46755d3cff