Many small changes: Changed keyboard mnemonics (e.g. Alt-F to open File menu) so they work on non-Mac platforms but are ignored on Macs; they did not work as expected on Macs.
Fixed a bug in "Open Javadoc API" for Java 6 classes. Since
the URLs in Java 6 Javadocs are absolute, the fully-qualified
class names were incorrect.

Changed the extension of external process files from
*.drjavaxml and *.drjavajar to *.djapp. The distinction
between XML and JAR file is made by examining the content.

To avoid confusion between the properties ${drjava.file}
(which contains the file name of the currently used DrJava
executable) and ${drjavajar.file} (which contains the name
of the *.drjavajar file that might enclose the *.drjavaxml
file), the latter was renamed to ${enclosing.djapp.file}.

Fixed a bug that made action properties eager, causing
the actions to be executed (!) when the values were
previewed. Now they are lazy again.

Fixed a bug in PredictiveInputFrame that didn't deactivate
the hourglass pane when the frame was closed using the [X]
or the red button, instead of using the "Cancel" button.
Removed the MainFrame.installModalWindowAdapter(Window w)
method that did not require to-front and close actions
because the use of that method lead to this dangerous
situation.

Gave buttons in PredictiveInputFrames the ability to have
tooltips. Now the tooltip for the "Fully Qualified" button
on the "Auto-Complete Word" dialog explains that it will
insert the fully-qualified class name of the selected class.

Fixed a bug in the External Process panel that didn't
display the command line to execute correctly after it
was evaluated.

Fixed a bug that did not let the DrJava survey window
pop up, even when there was no New Version popup (previously,
the decision was made to ever only display *one* popup
window on DrJava startup to avoid clutter).

Decided to write the timestamp of the last DrJava survey
window popup into the configuration even when the user
pressed "No, do not send information". That way, the user
isn't consistently bothered with the same question at every
start.



git-svn-id: file:///tmp/test-svn/trunk@4528 fe72c1cf-3628-48e9-8b72-1c46755d3cff