Docs: Important: migrated build process from docbook2html and docbook2pdf to the generic xsltproc tool.  This requires xsltproc on the path (/usr/bin on CSNet) and the variable DOCBOOK_XSL_HOME to point to the DocBook style sheets (/usr/share/sgml/docbook/xsl-stylesheets-1.69.1-5.1 on CSNet).  This migration was necessary because it's difficult to *find* docbook2html and docbook2pdf anymore.  The community seems to have moved to XML-processing tools.
(It would be nice to eliminate the dependency on xsltproc as well.  A variety of Java tools are available and could be bundled in lib/buildlib, but none of them seem to handle the DocBook styles correctly: the built-in Java APIs choke on the style definition; Xalan runs, but leaves the table-of-contents pages blank; Saxon doesn't support "chunking" -- converting a document to multiple HTML pages.)

The change in tools required some tweaking of image pointers: the only way to get consistent images of a reasonable size in both the HTML and PDF output seems to be by using a 'width="80%"' attribute (other percentages work too, of course).

Also made some changes in the developer docs:
- Some cleanup (related to recent changes in build scripts)
- Filled in stubs in a lot of "Supporting Technology" tools with some basic information.
- Put together an outline of milestones in the "History" section.
- Created a "Supporting Technology" section devoted to legacy tools/libraries (things we no longer use).
- Added stub sections to the "System Architecture" section.



git-svn-id: file:///tmp/test-svn/trunk@4712 fe72c1cf-3628-48e9-8b72-1c46755d3cff