Implemented task: Allow "Open Javadoc" for JUnit 3.8.2 https://sourceforge.net/pm/task.php?func=detailtask&project_task_id=145869&group_id=44253&group_project_id=54614
JUnit 3.8.2 Javadoc is now available for "Open Java API Javadoc"
and for auto-completion. The URL can be entered in the Javadoc
preference pane.

There is also a list of URLs ("Additional Javadoc URLs") where
additional Javadoc URLs can be entered and which will also be
parsed and used for "open Java API Javadoc" and auto-completion
(Example: add the URL http://drjava.org/javadoc/drjava to make
the DrJava documentation available).

Improved interface for configuring the auto-import classes for
the Interactions Pane. Now uses a list component instead of a
comma-separated string.



git-svn-id: file:///tmp/test-svn/trunk@4802 fe72c1cf-3628-48e9-8b72-1c46755d3cff