Fixes bug 2686460: Sometimes "Run" Doesn't Run, Only after Reset
I located the problem in the reset optimization. If the JVM is in the
FreshRunningState and a reset isn't forced, the JVM won't do
anything. We want to avoid the work of resetting the interpreter JVM,
but we do need the notification that the interpreter JVM is ready now.

The "java.lang.IllegalStateException: MainJVM is disposed" problems
remain.


git-svn-id: file:///tmp/test-svn/trunk@4812 fe72c1cf-3628-48e9-8b72-1c46755d3cff