This revision attempts to fix the consistent unit test failures we have been experiencing over the last six weeks.
When the JVM isn't in the restarting state anymore, start the JVM.
This gets it out of the fresh state and into the starting state.
Otherwise the interpreter may finish restarting but stay in the
fresh state, and asking the fresh state for the interpreter returns.

When we have disposed of the model, we remove all interaction
listeners to ensure we do not get any late notifications from the
interpreter JVM.

M    src/edu/rice/cs/drjava/model/GlobalModelTestCase.java
M    src/edu/rice/cs/drjava/model/repl/newjvm/MainJVM.java


git-svn-id: file:///tmp/test-svn/trunk@4832 fe72c1cf-3628-48e9-8b72-1c46755d3cff