Fixes bug 2738438: Saving File with class, but no Identifier throws Exception
Improves detection of empty files that trip up unit testing: If the
document doesn't produce a *.class file, then this will be considered
"out of sync" even though it isn't. We don't really have many options
here unless we actually parse the document.

As a heuristic, we check if there is any non-comment text that
includes the words "class", "interface" or "enum". If the file
doesn't, we presume it is empty and doesn't generate a class file;
therefore, it cannot ever be out of sync.

Improved error diagnostics if classes are out of sync before JUnit: The documents
in question are now actually displayed, and the error messages are more
precise in case compiling fails.

Moved MainFrame.setPopupLoc to Utilities.setPopupLoc. It doesn't
have anything to do with MainFrame.


git-svn-id: file:///tmp/test-svn/trunk@4872 fe72c1cf-3628-48e9-8b72-1c46755d3cff