More changes to address bug 2957033: Unable to view watch variables
We now display all watches correctly, and we do not allow method calls
or expressions.

We do not have good type information, though, because there is no way
to distinguish whether the interpreter returned a primitive or a boxed
type right now. Also, handling both local variables and fields would
be messy right now.

What we should do is also export the type of an interpretation result.



git-svn-id: file:///tmp/test-svn/trunk@5180 fe72c1cf-3628-48e9-8b72-1c46755d3cff