This implements feature request 970724: Simplify Running ACM Java Programs
There is a new option in Preferences/Interactions Pane
called "Smart Run ('java') Command". If enabled (default),
the "Run" button and the "java" command in the Interactions
Pane detects if the class (a) is an ACM Java Task Force
program (subclass of acm.program.Program) or (b) an applet.
If (a), then it runs the program's main method, but inserts
"code=MyClass" as argument 0. If (b), it starts the
applet viewer. Otherwise, the "java" command behaves just
as before.



git-svn-id: file:///tmp/test-svn/trunk@5249 fe72c1cf-3628-48e9-8b72-1c46755d3cff