Probably fixed bug 3037300: NullPointerException when Find in Missing File
When the change was made to allow users to abort a search that
encountered missing files, the getNextDocument was made to return null
to signal such an aborted search. That probably caused the
NullPointerException.

I wasn't able to reproduce the bug before I made this change,
though. This analysis is based purely on reading the source code.



git-svn-id: file:///tmp/test-svn/trunk@5352 fe72c1cf-3628-48e9-8b72-1c46755d3cff