Fixed bug 2976104: java.lang.StringIndexOutOfBoundsException: String index out of range: 94 at java.lang.String.substring(Unknown Source) at edu.rice.cs.drjava.platform.WindowsRegistry.nullTerminatedToString(WindowsRegistry.java:612)
According to the String Javadocs, "The length of the new String is a
function of the charset, and hence may not be equal to the length of
the byte array." 

We cannot use barr.length to index in the string, we
need to index in the array.

Cannot test this on a Chinese system, but I'm fairly certain this should
work.



git-svn-id: file:///tmp/test-svn/trunk@5353 fe72c1cf-3628-48e9-8b72-1c46755d3cff