Implemented feature request 3081586: Store Preferences within Projects
There now is an "Advanced" button on the "Project Properties"
dialog that takes the user to a secondary window. In that
window, the user can list (by name, using predictive input)
those preferences that should be saved in the project file.
By default, "Indent Level" and "Language Level" will be saved.

Note that for simplicity, the actual values are still
configured in the "Preferences" window (Edit menu). This
list only names the preferences whose values should be
stored in the project file when the project is saved, and
restored when the project is loaded again.



git-svn-id: file:///tmp/test-svn/trunk@5408 fe72c1cf-3628-48e9-8b72-1c46755d3cff