This revision performs three minor updates.
First, it updates the source to follow the JUnit 4 convention that
the Assert class resides in the package org.junit instead of
junit.framework.

Second, it deletes support for ConcJUnit which died after JUnit 6;
ConcJUnit did not work in Java 7 and it status in Java 8 was unknown
because no one reported using it.  At some point we should perhaps
resurrect it.  Finally, it

Finally, it adds the omitted jar files from the 2.12.3 Scala
distribution to the lib directory so that DrScala builds based
on the repository include all of these libraries.

The following files were modified, added or removed: